// src/redux/modalSlice/types.ts
export interface ModalState {
  isWellDoneDiaryModalOpen: boolean;
  isCreatedModalOpen: boolean;
  isAddDiaryModalOpen: boolean;
  isCreatedExercisesModalOpen: boolean;
  isModalOpen: boolean;
  showMobileMenu: boolean;
  isAddDiaryExercisesModalOpen: boolean;
  isWellDoneExercisesDiaryModal: boolean;
}